Refactor import paths for game and player management, update riddle database, and fix player controller CRUD operations
