+++ client/app.js
@@ -1,4 +1,4 @@
-import { runGame } from './client.services/gameManager.js';
+import { runGame } from './managers/gameManager.js';
 
 
 async function main() {
